Remove dead links
Weasyl Notifier’s Chrome extension is gone, and the other Chrome extensions are unavailable because of Manifest V3. I’m not sure if the Safari extension still works, but that’s a plaintext HTTP URL to a non-store page, and we’ll probably implement push notifications at some point anyway.